Fix test issues iwth installed packages
- Move the tests folder inside the package
- Move the download data module inside the test directory
- Force the tests to download the data if it does not exist
- Update the CI/Test workflow

+++ pysteps_nwp_importers/tests/download_test_data.py
@@ -7,9 +7,6 @@ DATA_DIR = Path(__file__).parent / "../tests/data"
 
 
 def download_test_data():
-
-    print("Downloading test data from github.")
-
     root_url = "https://github.com/pySTEPS/pysteps-data/raw/master/nwp/"
 
     files_to_download = (
@@ -24,8 +21,9 @@ def download_test_data():
         filename = Path(_file).name
         os.makedirs(DATA_DIR / subdir, exist_ok=True)
         dest_path = (DATA_DIR / subdir / filename).resolve()
-        print(file_url, "->", str(dest_path))
-        request.urlretrieve(file_url, dest_path)
+        if not dest_path.is_file():
+            print("Downloading ", file_url, "->", str(dest_path))
+            request.urlretrieve(file_url, dest_path)
 
 
 if __name__ == "__main__":
